+++ tests/testthat/test-AsylumResearch.R
@@ -1,11 +1,18 @@
 context("Asylum Research AFM image check")
 
 COMMON.INFO.ITEMS = 7
+filename = system.file("AR_20211011.ibw",package="nanoscopeAFM")
+
+test_that("AR: filefound", {
+  filename = system.file("AR_20211011.ibw",package="nanoscopeAFM")
+  dput(getwd())
+  expect_true(file.exists(filename))
+})
 
 
 test_that("AR: check channels ", {
-  filename = file.path('../../inst',dir(system.file(package = "nanoscopeAFM"),
-                                        pattern='ibw$', recursive = TRUE)[1])
+  filename = system.file("AR_20211011.ibw",package="nanoscopeAFM")
+
   h1 = read.AR_eofHeader.V2(filename)
   ch = .getChannelName(h1,1)
   expect_equal(.getChannelUnits(ch),"m")
@@ -20,8 +27,6 @@ test_that("AR: check channels ", {
 
 
 test_that("check AR image loads", {
-  filename = file.path('../../inst',dir(system.file(package = "nanoscopeAFM"),
-                                        pattern='ibw$', recursive = TRUE)[1])
   d = read.AR_file(filename,1)
   expect_equal(attr(d,'units'),"m")
   expect_equal(nrow(d),128*128)
@@ -33,8 +38,6 @@ test_that("check AR image loads", {
 
 
 test_that("check AR image loads for image 2", {
-  filename = file.path('../../inst',dir(system.file(package = "nanoscopeAFM"),
-                                        pattern='ibw$', recursive = TRUE)[1])
   d = read.AR_file(filename,2)
   expect_equal(attr(d,'units'),"m")
   expect_equal(nrow(d),128*128)
@@ -47,8 +50,6 @@ test_that("check AR image loads for image 2", {
 
 
 test_that("check AR image loads for image 3", {
-  filename = file.path('../../inst',dir(system.file(package = "nanoscopeAFM"),
-                                        pattern='ibw$', recursive = TRUE)[1])
   d = read.AR_file(filename,3)
   expect_equal(attr(d,'units'),"deg")
   expect_equal(nrow(d),128*128)
@@ -61,8 +62,6 @@ test_that("check AR image loads for image 3", {
 
 
 test_that("check AR image loads for image 4", {
-  filename = file.path('../../inst',dir(system.file(package = "nanoscopeAFM"),
-                                        pattern='ibw$', recursive = TRUE)[1])
   d = read.AR_file(filename,4)
   expect_equal(attr(d,'units'),"m")
   expect_equal(nrow(d),128*128)
@@ -75,8 +74,6 @@ test_that("check AR image loads for image 4", {
 
 
 test_that("use general AFM loading for AR image", {
-  filename = file.path('../../inst',dir(system.file(package = "nanoscopeAFM"),
-                                        pattern='ibw$', recursive = TRUE)[1])
   d = AFM.read(filename)
   expect_equal(nrow(d),128*128)
 })
@@ -84,8 +81,6 @@ test_that("use general AFM loading for AR image", {
 
 
 test_that("AR: check version 2 channel list ", {
-  filename = file.path('../../inst',dir(system.file(package = "nanoscopeAFM"),
-                                        pattern='ibw$', recursive = TRUE)[1])
   h1 = read.AR_eofHeader.V2(filename)
   expect_equal(length(h1), 13)
 })